extract commands out of dialog
