+++ package.json
@@ -39,7 +39,7 @@
         "@tanstack/react-query": "^5.51.11",
         "class-variance-authority": "^0.7.0",
         "clsx": "^2.1.1",
-        "cmdk": "^1.0.0",
+        "cmdk": "^1.0.4",
         "file-saver": "^2.0.5",
         "html-react-parser": "^5.1.12",
         "jszip": "^3.10.1",
